% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalanalysis.R
\name{bestmcc}
\alias{bestmcc}
\title{Returns the highest MCC among the all the MCCs calculated at cutoffs between 1\% to 100\% in sequence of 1\% .}
\usage{
bestmcc(actual, predicted)
}
\arguments{
\item{actual}{a vector of actual results from test set}

\item{predicted}{a vector of predicted results on test set}
}
\value{
Highest MCC value among all MCC values plotted against cutoff values of 0.1 to 1
}
\description{
Returns the highest MCC among the all the MCCs calculated at cutoffs between 1\% to 100\% in sequence of 1\% .
}
\keyword{Coefficient}
\keyword{Correlation}
\keyword{Matthews}
\keyword{confusion}
\keyword{evaluation,}
\keyword{matrix}
