% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkanalysis.R
\name{mapping_monkey}
\alias{mapping_monkey}
\title{Calculates network parameters from graph file and returns list of results}
\usage{
mapping_monkey(
  graphitem,
  mode = "all",
  norm = TRUE,
  direc = FALSE,
  loops = FALSE
)
}
\arguments{
\item{graphitem}{an Igraph item}

\item{mode}{used to calculate degree. defaults to "all"}

\item{norm}{defaults to TRUE. Used to control the parameter "normalized" in betweenness}

\item{direc}{defaults to FALSE. Used to control parameter direction in eigen_centrality, betweenness, diameter, farthest_vertices, get_diameter and assortativity.degree}

\item{loops}{defaults to FALSe. Used to control loop parameter in edge_density}
}
\value{
a list of results
}
\description{
Calculates network parameters from graph file and returns list of results
}
\keyword{analysis}
\keyword{network}
