% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_barscatterchart}
\alias{mod_barscatterchart}
\title{UI to produce a plotly output (horizontal bar chart + scatter plot) of height 200px}
\usage{
mod_barscatterchart(id)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding server function}
}
\value{
UI with a plotly output
}
\description{
UI to produce a plotly output (horizontal bar chart + scatter plot) of height 200px
}
\keyword{Horizontal}
\keyword{Multiple}
\keyword{Plotly,}
\keyword{Shiny,}
\keyword{and}
\keyword{bar}
\keyword{chart,}
\keyword{scatter}
\keyword{trace}
