% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_barscatterchartServer}
\alias{mod_barscatterchartServer}
\title{Server to produce a plotly output (line chart with fill i.e. area) of height 200px}
\usage{
mod_barscatterchartServer(id, df, title = "", name1 = "", name2 = "")
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding UI function}

\item{df}{a data frame with column 'name' (that contains factors), 'value1' and 'value2' columns that contain numbers .}

\item{title}{Text to add title to the graph}

\item{name1}{Text to add name to trace 1 (bar)}

\item{name2}{Text to add name to trace 2 (Scatter)}
}
\value{
UI and yaxis value of the click on the plot
}
\description{
Server to produce a plotly output (line chart with fill i.e. area) of height 200px
}
\keyword{Horizontal}
\keyword{Multiple}
\keyword{Plotly,}
\keyword{Shiny,}
\keyword{and}
\keyword{bar}
\keyword{chart,}
\keyword{scatter}
\keyword{trace}
