% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_gauge}
\alias{mod_gauge}
\title{UI to produce a plotly output (gauge chart) of height 150px}
\usage{
mod_gauge(id)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding server function}
}
\value{
UI with a plotly output
}
\description{
UI to produce a plotly output (gauge chart) of height 150px
}
\keyword{Chart,}
\keyword{Gauge}
\keyword{Plotly,}
\keyword{Shiny}
