% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_gaugeServer}
\alias{mod_gaugeServer}
\title{Server to produce a plotly output (line chart with fill i.e. area) of height 200px}
\usage{
mod_gaugeServer(
  id,
  value,
  title = "",
  lastvalue,
  maxval = 500,
  rangeval1 = 250,
  rangeval2 = 400,
  threshold = 490
)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding UI function}

\item{value}{an integer value to show the current value}

\item{title}{Text to add title to the graph}

\item{lastvalue}{Integer value that shows last value of the indicated value. This is used to calculate the delta vis-a-vis current value}

\item{maxval}{Integer to indicate maximum limit of the gauge}

\item{rangeval1}{Integer used to define the first segment of the gauge i.e. from 0 to this value is section 1}

\item{rangeval2}{Integer used to define the second segment. i.e. from rangeval1 to rangeval2 is second segment}

\item{threshold}{Integer to mark a threshold point in the gauge}
}
\value{
UI and with a gauge chart
}
\description{
Server to produce a plotly output (line chart with fill i.e. area) of height 200px
}
\keyword{Gauge}
\keyword{Plotly,}
\keyword{Shiny}
\keyword{chart,}
