% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_horizontalmainchart}
\alias{mod_horizontalmainchart}
\title{UI to produce a plotly output (horizontal bar chart) of height 200px}
\usage{
mod_horizontalmainchart(id)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding server function}
}
\value{
UI with a plotly output
}
\description{
UI to produce a plotly output (horizontal bar chart) of height 200px
}
\keyword{Horizontal}
\keyword{Plotly,}
\keyword{Shiny}
\keyword{bar}
\keyword{chart,}
