% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_horizontalmainchartServer}
\alias{mod_horizontalmainchartServer}
\title{Server to produce a plotly output (line chart with fill i.e. area) of height 200px}
\usage{
mod_horizontalmainchartServer(id, df, title = "")
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding UI function}

\item{df}{a data frame with column 'name' (that contains factors) and 'value' (that contains numbers).}

\item{title}{Text to add title to the graph}
}
\value{
UI and yaxis value of the click on the plot
}
\description{
Server to produce a plotly output (line chart with fill i.e. area) of height 200px
}
\keyword{Horizontal}
\keyword{Plotly,}
\keyword{Shiny}
\keyword{bar}
\keyword{chart,}
