% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_info}
\alias{mod_info}
\title{UI to produce infobox}
\usage{
mod_info(id)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding server function}
}
\value{
UI with a infobox output
}
\description{
UI to produce infobox
}
\keyword{Shiny}
\keyword{infobox,}
