% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_infoServer}
\alias{mod_infoServer}
\title{Server to produce infobox}
\usage{
mod_infoServer(id, value, title, colour, icon, subtitle = "")
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding UI function}

\item{value}{as the value to show in infobox}

\item{title}{as title to show in infobox}

\item{colour}{as colour of the infobox}

\item{icon}{name as font awesome icon name to be shown in the infobox}

\item{subtitle}{as text to show as subtitle in the infobox}

\item{yaxistitle}{defaults to empty. Can be used to add title to y axis of the poltly chart}
}
\value{
UI and x axis values (min and max) of selection on plot
}
\description{
Server to produce infobox
}
\keyword{Shiny}
\keyword{infobox,}
