% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_linechart}
\alias{mod_linechart}
\title{UI to produce a plotly output (line chart with fill i.e. area) of height 200px}
\usage{
mod_linechart(id)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding server function}
}
\value{
UI with a plotly output
}
\description{
UI to produce a plotly output (line chart with fill i.e. area) of height 200px
}
\keyword{Area}
\keyword{Chart,}
\keyword{Line}
\keyword{Plotly,}
\keyword{Shiny}
