% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_linechartnofill}
\alias{mod_linechartnofill}
\title{UI to produce a plotly output (line chart without fill) of height 200px}
\usage{
mod_linechartnofill(id)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding server function}
}
\value{
UI with a plotly output
}
\description{
UI to produce a plotly output (line chart without fill) of height 200px
}
\keyword{Line}
\keyword{Plotly,}
\keyword{Shiny,}
\keyword{chart,}
