% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_linechartnofillServer}
\alias{mod_linechartnofillServer}
\title{Server to produce a plotly output (line chart without fill) of height 200px}
\usage{
mod_linechartnofillServer(id, df, yaxistitle = "")
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding UI function}

\item{df}{a data frame that has 'datetime' as a columns of date/datetime and 'value' as a column of numbers corresponding to value}

\item{yaxistitle}{defaults to empty. Can be used to add title to y axis of the poltly chart}
}
\value{
UI and x axis values (min and max) of selection on plot
}
\description{
Server to produce a plotly output (line chart without fill) of height 200px
}
\keyword{Plotly,line}
\keyword{Shiny}
\keyword{chart,}
