% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_linechartServer}
\alias{mod_linechartServer}
\title{Server to produce a plotly output (line chart with fill i.e. area) of height 200px}
\usage{
mod_linechartServer(id, df, yaxistitle = "", fill = "tozeroy")
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding UI function}

\item{df}{a data frame that has 'datetime' as a columns of date/datetime and 'value' as a column of numbers corresponding to value}

\item{yaxistitle}{defaults to empty. Can be used to add title to y axis of the poltly chart.}

\item{fill}{defaults to "tozeroy". Defines the colour of the filled area}
}
\value{
UI and x axis values (min and max) of selection on plot
}
\description{
Server to produce a plotly output (line chart with fill i.e. area) of height 200px
}
\keyword{Area}
\keyword{Chart,}
\keyword{Line}
\keyword{Plotly}
\keyword{Plotly,}
\keyword{Shiny,}
\keyword{event}
