% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_mainchart}
\alias{mod_mainchart}
\title{UI to produce a plotly output (bar chart) of height 200px}
\usage{
mod_mainchart(id)
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding server function}
}
\value{
UI with a plotly output
}
\description{
UI to produce a plotly output (bar chart) of height 200px
}
\keyword{Bar}
\keyword{Chart,}
\keyword{Plotly,}
\keyword{Shiny}
