% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModules.R
\name{mod_mainchartServer}
\alias{mod_mainchartServer}
\title{Server to produce a plotly output (bar chart) of height 200px}
\usage{
mod_mainchartServer(id, df, title = "")
}
\arguments{
\item{id}{an ID wrapped in double quotes that matches the id of the corresponding UI function}

\item{df}{a data frame that has 'name' as a columns of factors and 'value' as a column of numbers corresponding to value}

\item{title}{defaults to empty. Can be used to add title to the poltly chart.}
}
\value{
UI and x axis value of click on the plotly chart
}
\description{
Server to produce a plotly output (bar chart) of height 200px
}
\keyword{Bar}
\keyword{Chart,}
\keyword{Plotly}
\keyword{Plotly,}
\keyword{Shiny,}
\keyword{event}
