% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalanalysis.R
\name{bestmccplot}
\alias{bestmccplot}
\title{Plot Matthews Correlation Coefficient (MCC) against cut off.}
\usage{
bestmccplot(actual, predicted)
}
\arguments{
\item{actual}{a vector of actual results from test set}

\item{predicted}{a vector of predicted results on test set}
}
\value{
a plot with cutoff in x axis and MCC in y axis
}
\description{
Plot Matthews Correlation Coefficient (MCC) against cut off.
}
\keyword{Coefficient}
\keyword{Correlation}
\keyword{Matthews}
\keyword{confusion}
\keyword{evaluation,}
\keyword{matrix}
