% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juggling_jaguar.R
\name{juggling_jaguar}
\alias{juggling_jaguar}
\title{Generates combination of 2 columns from a data frame}
\usage{
juggling_jaguar(data.df, rm.na = TRUE)
}
\arguments{
\item{data.df}{a data frame of more than 2 columns}

\item{rm.na}{is set to true by default. It removes any rows with NA}
}
\value{
a data frame with supplied number of column, which is created from the sets by binding
}
\description{
Generates combination of 2 columns from a data frame
}
\keyword{analysis}
\keyword{network}
